requirements.txt
pandas>=2.2.0
python-dotenv>=1.0.1
requests>=2.32.0
beautifulsoup4>=4.12.0
scikit-learn>=1.5.0
